Replace old django interface by the SPA
Remove old files, test, urls, etc
