+++ bower.json
@@ -1,5 +1,5 @@
 {
-  "name": "fotuto",
+  "name": "Fotuto",
   "version": "1.0a",
   "homepage": "https://github.com/cesarcruz/fotuto",
   "authors": [
@@ -8,9 +8,6 @@
   "description": "Fotuto is a web SCADA to monitor variables behaviour in real time.",
   "private": false,
   "dependencies": {
-    "bootstrap": "~3.3.2",
-    "jquery": "~2.1.3",
-    "fontawesome": "~4.2.0",
     "angular": "^1.5.3",
     "angular-mocks": "^1.5.3",
     "angular-route": "^1.5.3",
@@ -18,6 +15,7 @@
     "angular-material": "^1.0.7",
     "mdi": "^1.5.54",
     "angular-sanitize": "^1.5.3",
+    "material-design-icons": "^2.2.3",
     "d3": "^3.5.16"
   }
 }
