+++ fotuto/urls.py
@@ -4,8 +4,10 @@ from django.contrib import admin
 from django.utils.text import slugify
 from django.views.generic.base import TemplateView
 from rest_framework.authtoken import views
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+from rest_framework.reverse import reverse as drf_reverse
 
-from windows.views import api_root
 try:
     from settings_local import ADDITIONAL_APPS
 except ImportError:
@@ -15,13 +17,19 @@ admin.site.site_title = "Fotuto Admin Site"
 admin.site.site_header = "Fotuto administration"
 admin.site.index_title = "Fotuto administration"
 
+@api_view(['GET'])
+def api_root(request, format=None):
+    return Response({
+        'windows': drf_reverse('window-list', request=request, format=format),
+        'devices': drf_reverse('device-list', request=request, format=format),
+        'users': drf_reverse('user-list', request=request, format=format),
+    })
+
 urlpatterns = [
-    url(r'^api/token/', views.obtain_auth_token, name='api-token'),
-    url(r'^api/$', api_root),
-    url(r'^api/', include('rest_framework.urls', namespace='rest_framework')),
-    url(r'^', include('django.contrib.auth.urls')),
     url(r'^$', TemplateView.as_view(template_name='spa/index.html'), {'addons': ADDITIONAL_APPS}),
-    # url(r'^', include('django.contrib.auth.urls')),
+    url(r'^api/$', api_root),
+    url(r'^api/token/', views.obtain_auth_token, name='api-token'),
+    url(r'^api/', include('rest_framework.urls', namespace='rest_framework')),
     url(r'^', include('operators.urls')),
     url(r'^', include('windows.urls')),
     url(r'^', include('mimics.urls')),
