--- functional_tests/test_user_visitor.py
@@ -1,6 +1,0 @@
-from .base import FunctionalTest
-
-
-class VisitorTest(FunctionalTest):
-    def test_visitors_cant_view_login_form_only(self):
-        pass
