+++ static/spa/index.html
@@ -4,6 +4,7 @@
 	<title ng-bind="'Fotuto | ' + title"></title>
 	<link href="{{ STATIC_URL }}bower_components/angular-material/angular-material.css" rel="stylesheet"/>
 	<link href="{{ STATIC_URL }}bower_components/mdi/css/materialdesignicons.min.css" rel="stylesheet"/>
+	<link href="{{ STATIC_URL }}bower_components/material-design-icons/iconfont/material-icons.css" rel="stylesheet"/>
 	<link href="{{ STATIC_URL }}spa/assets/fotuto-app.css" rel="stylesheet"/>
 	{# TODO: Including a file doesn't solve the problem of duplicated or using diferent version of the files #}
 	{% for addon in addons %}
