--- templates/base.html
@@ -1,104 +1,0 @@
-{% load bootstrap3 %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<title>Fotuto | {% block title %}Home{% endblock %}</title>
-	<link href="{{ STATIC_URL }}bower_components/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" media="screen">
-	<link href="{{ STATIC_URL }}bower_components/fontawesome/css/font-awesome.min.css" rel="stylesheet">
-	<!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
-	<!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-	<!--[if lt IE 9]>
-		<script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-		<script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
-	<link href="{{ STATIC_URL }}css/styles.css" rel="stylesheet">
-</head>
-<body>
-<nav class="navbar navbar-inverse navbar-fixed-top" role="navigation">
-	<div class="container-fluid">
-		<!-- Brand and toggle get grouped for better mobile display -->
-		<div class="navbar-header">
-			<button type="button" class="navbar-toggle collapsed" data-toggle="collapse"
-				data-target="#main_menu">
-				<span class="sr-only">Toggle navigation</span>
-				{% bootstrap_icon ' fa fa-bars' %}
-			</button>
-			<a class="navbar-brand" href="/">Fotuto</a>
-		</div>
-
-		<!-- Collect the nav links, forms, and other content for toggling -->
-		<div id="main_menu" class="collapse navbar-collapse">
-			<ul class="nav navbar-nav">
-				<li class="mimics dropdown active">
-					<a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false">
-						<i class="fa fa-dashboard fa-fw"></i> Dashboards <span class="caret"></span></a>
-
-					<ul class="dropdown-menu" role="menu">
-						<!-- TODO: Use nested submenus for manage and add -->
-						<li class="dropdown-header">
-							{% if perms.vars.change_var %}
-								<i class="fa fa-gears fa-fw"></i> Manage
-							{% else %}
-								<i class="fa fa-list-ul fa-fw"></i> List
-							{% endif %}
-						</li>
-						<li class="disabled"><a href="{% url 'mimic_list' %}"><i class="fa fa-dashboard fa-fw"></i> Mimics</a></li>
-						<li><a href="{% url 'window_list' %}"><i class="fa fa-desktop fa-fw"></i> Windows</a></li>
-						<li><a href="{% url 'device_list' %}"><i class="fa fa-hdd-o fa-fw"></i> Devices</a></li>
-						<li><a href="{% url 'var_list' %}"><i class="fa fa-crosshairs fa-fw"></i> Vars</a></li>
-						{% if perms.vars.add_var %}
-							<li class="divider"></li>
-							<li class="dropdown-header"><i class="fa fa-plus-square fa-fw"></i> Add</li>
-							<li class="disabled"><a href="{% url 'mimic_add' %}"><i class="fa fa-dashboard fa-fw"></i> Mimic</a></li>
-							<li><a href="{% url 'window_add' %}"><i class="fa fa-desktop fa-fw"></i> Window</a></li>
-							<li><a href="{% url 'device_add' %}"><i class="fa fa-hdd-o fa-fw"></i> Device</a></li>
-							<li><a href="{% url 'var_add' %}"><i class="fa fa-crosshairs fa-fw"></i> Var</a></li>
-						{% endif %}
-						<!-- TODO: Display windows -->
-					</ul>
-				</li>
-				<li><a href="#"><i class="fa fa-area-chart fa-fw"></i> History charts</a></li>
-			</ul>
-			<div id="navbar-tray" class="navbar-right">
-				{% if user.is_authenticated %}
-					<p class="navbar-text">Welcome <a href="#" class="navbar-link">{{ user.username }}</a>
-						<a href="{% url 'logout' %}">Logout</a></p>
-				{% else %}
-					<a class="btn btn-default navbar-btn btn-signin" href="{% url 'login' %}">Sign In</a>
-				{% endif %}
-			</div>
-		</div>
-		<!-- /.navbar-collapse -->
-	</div>
-	<!-- /.container-fluid -->
-</nav>
-
-<div id="breadcrumbs">
-	<ol class="breadcrumb">
-		<li><a href="/">Home</a></li>
-		{% block breadcrums %}{% endblock %}
-	</ol>
-</div>
-<div class="container-fluid">
-	<div class="page-header">
-		<div class="pull-right btn-group">{% block toolsbar %}{% endblock %}</div>
-		<h1>{% block header_text %}Home{% endblock %}</h1>
-	</div>
-
-	{% if messages %}
-		{% for message in messages %}
-			<div class="alert alert-{{ message.tags }}"><i
-				class="fa fa-{% if message.tags == 'success' %}check{% elif message.tags == 'warning' %}exclamation{% elif message.tags == 'danger' %}minus{% else %}{{ message.tags }}{% endif %}-circle"></i> {{ message }}
-			</div>
-		{% endfor %}
-	{% endif %}
-	{% block messages %}{% endblock %}
-	{% block content %}{% endblock %}
-</div>
-<script src="{{ STATIC_URL }}bower_components/jquery/dist/jquery.min.js"></script>
-<script src="{{ STATIC_URL }}bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
-</body>
-</html>
