--- templates/base_form.html
@@ -1,17 +1,0 @@
-{% extends 'base.html' %}
-
-{% load bootstrap3 %}
-{# TODO: Add form.media resources and form.errors messages #}
-
-{% block content %}
-	<form method="post" action="{% block form_action %}{% endblock %}">
-		{% csrf_token %}
-		{% bootstrap_form form %}
-		{% buttons %}
-            <button type="submit" class="btn btn-primary {% block button_class %}{% endblock %}">
-                {% block button_text %}{% bootstrap_icon " fa fa-save" %} Save{% endblock %}
-            </button>
-        {% endbuttons %}
-	{% block form_append %}{% endblock %}
-	</form>
-{% endblock %}
