--- templates/registration/logged_out.html
@@ -1,9 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Logout{% endblock %}
-
-{% block header_text %}Logout{% endblock %}
-
-{% block breadcrums %}<li class="active">Logout</li>{% endblock %}
-
-{% block content %}You has been logged out successfully.{% endblock %}
