--- templates/registration/login.html
@@ -1,15 +1,0 @@
-{% extends "base_form.html" %}
-
-{% block title %}Sign In{% endblock %}
-
-{% block header_text %}Sign In{% endblock %}
-
-{% block breadcrums %}<li class="active">Sign In</li>{% endblock %}
-
-{% block form_action %}{% url 'django.contrib.auth.views.login' %}{% endblock %}
-
-{% block form_append %}<input type="hidden" name="next" value="{{ next }}" />{% endblock %}
-
-{% block button_class %}btn-login{% endblock %}
-
-{% block button_text %}Login{% endblock %}
