--- vars/templates/vars/device_form.html
@@ -1,10 +1,0 @@
-{% extends 'base_form.html' %}
-
-{% block title %}Add Device{% endblock %}
-
-{% block header_text %}Add Device{% endblock %}
-
-{% block breadcrums %}
-	<li><a href="{% url 'device_list' %}">Devices</a></li>
-	<li class="active">Add new</li>
-{% endblock %}
