--- vars/templates/vars/device_list.html
@@ -1,43 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Devices{% endblock %}
-
-{% block header_text %}<i class="fa fa-hdd-o"></i> Devices{% endblock %}
-
-{% block breadcrums %}
-	<li class="active">Devices</li>
-{% endblock %}
-
-{% block toolsbar %}
-	<a href="{% url 'device_add' %}" class="btn btn-primary"><i class="fa fa-plus-square"></i> Add Device</a>
-{% endblock %}
-
-{% block content %}
-	<table class="table">
-		<tr>
-			<th>Name</th>
-			<th>Address</th>
-			<th>Visible</th>
-			<th>Model</th>
-			<th>Description</th>
-			<th>Actions</th>
-		</tr>
-		{% for device in device_list %}
-			<tr class="{{ device.active|yesno:',text-muted' }}">
-				<td>{{ device.name }}</td>
-				<td>{{ device.address }}</td>
-				<td><i class="text-{{ device.active|yesno:'success,danger' }} fa fa-eye{{ device.active|yesno:',-slash' }}" title="{{ device.active|yesno:"Active,Inactive" }}"></i></td>
-				<td>{{ device.model }}</td>
-				<td>{{ device.description }}</td>
-				<td class="btn-group-xs">
-					<a href="#" class="btn btn-default" title="Edit"><i class="fa fa-pencil"></i></a>
-					<a href="#" class="btn btn-danger" title="Delete"><i class="fa fa-trash"></i></a>
-				</td>
-			</tr>
-		{% empty %}
-			<tr>
-				<td colspan="6" class="alert alert-warning"><i class="fa fa-exclamation-circle"></i> No device found.</td>
-			</tr>
-		{% endfor %}
-	</table>
-{% endblock %}
