--- vars/templates/vars/var_form.html
@@ -1,15 +1,0 @@
-{% extends 'base_form.html' %}
-
-{% block title %}Add Variable{% endblock %}
-
-{% block header_text %}Add Variable{% endblock %}
-
-{% block breadcrums %}
-	<li><a href="/vars/">Variables</a></li>
-	<li class="active">Add new</li>
-{% endblock %}
-
-{% block toolsbar %}
-	<a href="{% url 'device_add' %}" class="btn btn-primary"><i class="fa fa-plus-square"></i> Add
-		device</a>
-{% endblock %}
