--- vars/templates/vars/var_list.html
@@ -1,46 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Variables{% endblock %}
-
-{% block header_text %}<i class="fa fa-crosshairs"></i> Variables{% endblock %}
-
-{% block breadcrums %}
-	<li class="active">Variables</li>
-{% endblock %}
-
-{% block toolsbar %}
-	<a href="{% url 'var_add' %}" class="btn btn-primary"><i class="fa fa-plus-square"></i> Add Variable</a>
-{% endblock %}
-
-{% block content %}
-	<table class="table">
-		<tr>
-			<th>Name</th>
-			<th>Device</th>
-			<th>Visible</th>
-			<th>Type</th>
-			<th>Units</th>
-			<th>Description</th>
-			<th>Actions</th>
-		</tr>
-		{% for var in var_list %}
-			<tr class="{{ var.active|yesno:',text-muted' }}">
-				<td>{{ var.name }}</td>
-				<td>{{ var.device }}</td>
-				<td><i class="text-{{ var.active|yesno:'success,danger' }} fa fa-eye{{ var.active|yesno:',-slash' }}" title="{{ var.active|yesno:"Active,Inactive" }}"></i>
-				</td>
-				<td>{{ var.get_var_type_display }}</td>
-				<td>{{ var.units }}</td>
-				<td>{{ var.description }}</td>
-				<td class="btn-group-xs">
-					<a href="#" class="btn btn-default" title="Edit"><i class="fa fa-pencil"></i></a>
-					<a href="#" class="btn btn-danger" title="Delete"><i class="fa fa-trash"></i></a>
-				</td>
-			</tr>
-		{% empty %}
-			<tr>
-				<td colspan="2" class="alert alert-warning"><i class="fa fa-exclamation-circle"></i> No variable found.</td>
-			</tr>
-		{% endfor %}
-	</table>
-{% endblock %}
