--- windows/templates/windows/window_form.html
@@ -1,10 +1,0 @@
-{% extends 'base_form.html' %}
-
-{% block title %}Add Window{% endblock %}
-
-{% block header_text %}Add Window{% endblock %}
-
-{% block breadcrums %}
-	<li><a href="{% url 'window_list' %}">Windows</a></li>
-	<li class="active">Add new</li>
-{% endblock %}
