--- windows/templates/windows/window_list.html
@@ -1,40 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Windows{% endblock %}
-
-{% block header_text %}<i class="fa fa-desktop"></i> Windows{% endblock %}
-
-{% block breadcrums %}
-	<li class="active">Windows</li>
-{% endblock %}
-
-{% block toolsbar %}
-	<a href="{% url 'window_add' %}" class="btn btn-primary"><i class="fa fa-plus-square"></i> Add Window</a>
-{% endblock %}
-
-{% block content %}
-	<table class="table">
-		<tr>
-			<th>Title</th>
-			<th>Description</th>
-			<th>Actions</th>
-		</tr>
-		{% for window in window_list %}
-			<tr>
-				<td><a href="{{ window.get_absolute_url }}">{{ window.title }}</a></td>
-				<td>{{ window.description }}</td>
-				<td class="btn-group-xs">
-					<a href="{{ window.get_absolute_url }}" class="btn btn-default view" title="View">View</a>
-					<a href="{% url 'mimic_manage_window' window.slug %}" class="btn btn-default manage-mimics"
-						title="Manage Mimics"><i class="fa fa-dashboard"></i></a>
-					<a href="#" class="btn btn-default" title="Edit"><i class="fa fa-pencil"></i></a>
-					<a href="#" class="btn btn-danger" title="Delete"><i class="fa fa-trash"></i></a>
-				</td>
-			</tr>
-		{% empty %}
-			<tr>
-				<td colspan="3" class="alert alert-warning"><i class="fa fa-exclamation-circle"></i> No window found.</td>
-			</tr>
-		{% endfor %}
-	</table>
-{% endblock %}
