+++ windows/tests/test_serializers.py
@@ -1,10 +1,9 @@
 from unittest import TestCase
 
 from django.contrib.auth import get_user_model
-from rest_framework.test import APIRequestFactory, force_authenticate
+from rest_framework.test import APIRequestFactory
 
 from windows.serializers import WindowSerializer
-from windows.views import api_root
 
 User = get_user_model()
 
@@ -37,18 +36,3 @@ class WindowSerializerTestCase(TestCase):
             'mimics': 'http://testserver/api/mimics/?window=%s' % window.pk,
             'vars': 'http://testserver/api/vars/?mimic__window=%s' % window.pk,
         }}, serializer.data,)
-
-
-class APIRootURLTestCase(TestCase):
-    def test_api_root_return_correct_urls(self):
-        factory = APIRequestFactory()
-        user = User.objects.create_user(username='maximo')
-        request = factory.get('/api/groups/', )
-        force_authenticate(request, user=user)
-        response = api_root(request=request)
-        expected_json = {
-            'windows': 'http://testserver/api/windows/',
-            'devices': 'http://testserver/api/devices/',
-            'users': 'http://testserver/api/users/',
-        }
-        self.assertDictEqual(response.data, expected_json)
